+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - build.sh: add template context
 - add a --build-script-only flag, to render the build.sh script
 - add a --tag-name option, to support automatic asset upload
+- add support to custom build template scripts